pyproject.toml
[project]
name = "nimui"
version = "0.1.0"
description = "A simple script using NVIDIA API with requests and optional dotenv"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
license = "MIT"
dependencies = [
    "requests>=2.32.4",
    "python-dotenv>=1.2.2",
    "pathspec>=0.12.1",
    "rich>=13.9.4"
]

[project.scripts]
chat = "nimui.cli:main"

[build-system]
requires = ["setuptools>=77.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["nimui*"]

[tool.setuptools.package-data]
nimui = ["models.json"]